+++ apps/dragonfly_browser/lib/src/screens/browser/blocs/browser_cubit.dart
@@ -1,5 +1,5 @@
 import 'package:dragonfly/main.dart';
-import 'package:dragonfly_navigation/dragonfly_navigation.dart';
+import 'package:dragonfly_engine/dragonfly_engine.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class BrowserCubit extends Cubit<Browser> {
