+++ apps/dragonfly_browser/lib/src/screens/browser/browser_theme.dart
@@ -1,4 +1,4 @@
-import 'package:dragonfly_navigation/dragonfly_navigation.dart';
+import 'package:dragonfly_engine/dragonfly_engine.dart';
 import 'package:flutter/material.dart';
 
 class CSSOMProvider extends InheritedWidget {
