+++ apps/dragonfly_browser/lib/src/screens/browser/pages/file_explorer_page.dart
@@ -5,7 +5,7 @@ import 'package:collection/collection.dart';
 import 'package:dragonfly/src/constants/file_constants.dart';
 import 'package:dragonfly/src/screens/browser/blocs/browser_cubit.dart';
 import 'package:dragonfly/src/screens/browser/pages/cubit/file_explorer_cubit.dart';
-import 'package:dragonfly_navigation/dragonfly_navigation.dart';
+import 'package:dragonfly_engine/dragonfly_engine.dart';
 import 'package:flutter/material.dart' hide Tab;
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:url_launcher/url_launcher.dart';
