+++ apps/dragonfly_browser/lib/src/screens/browser/pages/json/json_screen.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:dragonfly_navigation/dragonfly_navigation.dart';
+import 'package:dragonfly_engine/dragonfly_engine.dart';
 import 'package:flutter/material.dart';
 
 class JsonScreen extends StatefulWidget {
