+++ apps/dragonfly_browser/lib/src/screens/scaffold/mobile/mobile_scaffold.dart
@@ -1,10 +1,9 @@
 import 'package:dragonfly/src/screens/browser/blocs/browser_cubit.dart';
 import 'package:dragonfly/src/screens/browser/browser_screen.dart';
 import 'package:dragonfly/src/screens/scaffold/browser_scaffold.dart';
-import 'package:dragonfly_navigation/dragonfly_navigation.dart';
+import 'package:dragonfly_engine/dragonfly_engine.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
 
 class MobileScaffold extends StatelessWidget {
   const MobileScaffold({super.key});
