+++ apps/dragonfly_browser/lib/src/screens/scaffold/widgets/favicon_icon.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:dragonfly_navigation/dragonfly_navigation.dart';
+import 'package:dragonfly_engine/dragonfly_engine.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_svg/flutter_svg.dart';
