+++ apps/dragonfly_browser/pubspec.yaml
@@ -19,7 +19,7 @@ dependencies:
   syntax_highlight: ^0.4.0
   window_manager: ^0.4.2
   dragonfly_browservault: ^0.0.1
-  dragonfly_navigation: ^0.0.1
+  dragonfly_engine: ^0.0.1
   intl: ^0.19.0
   html: ^0.15.4
   url_launcher: ^6.3.0
