+++ packages/dragonfly_engine/lib/src/files/cache_file.dart
@@ -2,7 +2,7 @@
 import 'dart:io';
 import 'dart:typed_data';
 import 'package:dragonfly_browservault/dragonfly_browservault.dart';
-import 'package:dragonfly_navigation/dragonfly_navigation.dart';
+import 'package:dragonfly_engine/dragonfly_engine.dart';
 import 'package:path/path.dart' as p;
 import 'package:uuid/v4.dart';
