+++ packages/dragonfly_engine/lib/src/files/favicon.dart
