+++ packages/dragonfly_engine/lib/src/pages/a_page.dart
@@ -1,4 +1,4 @@
-import 'package:dragonfly_navigation/dragonfly_navigation.dart';
+import 'package:dragonfly_engine/dragonfly_engine.dart';
 import 'package:html/dom.dart';
 import 'package:uuid/uuid.dart';
 import 'package:path/path.dart' as p;
