+++ packages/dragonfly_engine/pubspec.yaml
@@ -1,5 +1,5 @@
-name: dragonfly_navigation
-description: A starting point for Dart libraries or applications.
+name: dragonfly_engine
+description: A web engine in Dart
 version: 0.0.1
 # repository: https://github.com/my_org/my_repo
