+++ packages/dragonfly_engine/test/network_tracker_test.dart
@@ -1,4 +1,4 @@
-import 'package:dragonfly_navigation/src/utils/network_tracker.dart';
+import 'package:dragonfly_engine/src/utils/network_tracker.dart';
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
 import 'package:http/http.dart' as http;
